Redid block renderering to allow for meshes, changed blocks and items to require id in constructor, fixed mushrooms breaking at the wrong time and added parts to the torch and lanturn

+++ Assets/VoxelEngine/Blocks/BlockChest.cs
@@ -10,7 +10,7 @@ namespace VoxelEngine.Blocks {
 
     public class BlockChest : BlockTileEntity {
 
-        public BlockChest(int id) : base(id) { }
+        public BlockChest(byte id) : base(id) { }
 
         public override void onRightClick(World world, EntityPlayer player, BlockPos pos, byte meta) {
             player.openContainer(References.list.containerChest, ((TileEntityChest)world.getTileEntity(pos)).chestData);
