+++ Assets/VoxelEngine/Blocks/BlockGrass.cs
@@ -5,7 +5,7 @@ namespace VoxelEngine.Blocks {
 
     public class BlockGrass : Block {
 
-        public BlockGrass(int id) : base(id) {}
+        public BlockGrass(byte id) : base(id) {}
 
         public override void onRandomTick(World world, int x, int y, int z, byte meta, int tickSeed) {
             //int i = (tickSeed) & 0x03;
