+++ Assets/VoxelEngine/Blocks/BlockOre.cs
@@ -9,7 +9,7 @@ namespace VoxelEngine.Blocks {
         private Item droppedItem;
         private int textureY;
 
-        public BlockOre(int id, Item drop, int textureY) : base(id) {
+        public BlockOre(byte id, Item drop, int textureY) : base(id) {
             this.droppedItem = drop;
             this.textureY = textureY;
         }
