+++ Assets/VoxelEngine/Blocks/BlockWood.cs
@@ -4,7 +4,7 @@ namespace VoxelEngine.Blocks {
 
     public class BlockWood : Block {
 
-        public BlockWood(int id) : base(id) { }
+        public BlockWood(byte id) : base(id) { }
 
         public override TexturePos getTexturePos(Direction direction, byte meta) {
             TexturePos tile = new TexturePos(1, 1);
